Add dual serdes for FullyQualifiedEntityName to allow for Json Object and JsString deserialization.
